chore: create a ValueRangeSlider package with the ValueRangeSlider object
